+++ changelog.rst
@@ -1,3 +1,11 @@
+Upcoming
+========
+
+Features:
+---------
+
+* Add support for `\T` command to change format output. (Thanks: `Frederic Aoustin`_).
+
 1.8.2
 =====
 
@@ -766,3 +774,4 @@ Improvements:
 .. _`Marcin Sztolcman`: https://github.com/msztolcman
 .. _`Isank`: https://github.com/isank
 .. _`Bojan Delić`: https://github.com/delicb
+.. _`Frederic Aoustin`: https://github.com/fraoustin